Refactor libpsl to use GetPwUid()
