+++ src/libpsl-native/src/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(psl-native SHARED
   getstat.cpp
+  getpwuid.cpp
   getfileowner.cpp
   getcurrentprocessorid.cpp
   getusername.cpp
