+++ src/libpsl-native/src/getpwuid.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <sys/types.h>
+#include "pal.h"
+
+PAL_BEGIN_EXTERNC
+
+char* GetPwUid(uid_t uid);
+
+PAL_END_EXTERNC
